refactor: 832 - rename folders and files in mentorship
