+++ src/widgets/mentorship-after-register/index.ts
@@ -1,0 +1,1 @@
+export { MentorshipAfterRegister } from './ui/mentorship-after-register';
