+++ src/widgets/mentorship-details/details/mentorship-details.module.scss
